Revert "notification dropdown (#11)"
This reverts commit e6b5302eabe1127a008189d0a16f94a1fdeea8fe.
