+++ index.html
@@ -15,8 +15,6 @@
     <meta name="twitter:card" content="summary_large_image" />
     <meta name="twitter:site" content="@lovable_dev" />
     <meta name="twitter:image" content="https://lovable.dev/opengraph-image-p98pqg.png" />
-
-    <link rel="icon" type="image/svg+xml" href="/linkedin.svg" />
   </head>
 
   <body>
