+++ src/App.tsx
@@ -5,13 +5,8 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Index from "./pages/Index";
 import HackathonDetails from "./pages/HackathonDetails";
-import ProgressReport from "./pages/ProgressReport";
 import NotFound from "./pages/NotFound";
 import QuestPage from "./pages/DailyQuest";
-import MCQRound from "./components/MCQRound";
-import SubmissionRound from "./components/SubmissionRound";
-import WeeklyWisdomQuizPage from "./pages/WeeklyWisdomQuizPage";
-import RedeemXP from "./pages/RedeemXP";
 
 const queryClient = new QueryClient();
 
@@ -25,11 +20,6 @@ const App = () => (
           <Route path="/" element={<Index />} />
           <Route path="/quest/:questId" element={<QuestPage />} />
           <Route path="/hackathon/:id" element={<HackathonDetails />} />
-          <Route path="/hackathon/:id/mcq" element={<MCQRound />} />
-          <Route path="/hackathon/:id/submission" element={<SubmissionRound />} />
-          <Route path="/progress" element={<ProgressReport />} />
-          <Route path="/weekly-wisdom-quiz" element={<WeeklyWisdomQuizPage />} />
-          <Route path="/redeem-xp" element={<RedeemXP />} />
           {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
           <Route path="*" element={<NotFound />} />
         </Routes>
