+++ src/pages/Index.tsx
@@ -3,7 +3,6 @@ import SkillGraph from "@/components/SkillGraph";
 import DailyQuest from "@/components/DailyQuest";
 import GamificationStats from "@/components/GamificationStats";
 import WeeklyHackathon from "@/components/WeeklyHackathon";
-import WeeklyWisdomQuiz from "@/components/WeeklyWisdomQuiz";
 import LlamaBot from "@/components/LlamaBot";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
@@ -16,14 +15,10 @@ import {
   Calendar,
   BookOpen,
   Users,
-  Lightbulb,
-  Star
+  Lightbulb
 } from "lucide-react";
-import { useNavigate } from "react-router-dom";
 
 const Index = () => {
-  const navigate = useNavigate();
-  
   const upcomingEvents = [
     // { type: "Quest", title: "Debug Challenge", time: "Tomorrow" },
     // { type: "Hackathon", title: "Mobile App Challenge", time: "This Weekend" },
@@ -35,14 +30,6 @@ const Index = () => {
     "You're on track to reach Level 13 by next month",
     "Consider exploring System Design - it's trending in your network"
   ];
-
-  const handleViewProgressReport = () => {
-    navigate('/progress');
-  };
-
-  const handleRedeemXP = () => {
-    navigate('/redeem-xp');
-  };
 
   return (
     <div className="min-h-screen bg-background">
@@ -64,20 +51,11 @@ const Index = () => {
         <div className="grid grid-cols-1 xl:grid-cols-3 gap-6 mb-8">
           {/* Left Column - Main Content */}
           <div className="xl:col-span-2 space-y-6">
-            {/* Weekly Wisdom Quiz (now first) */}
-            <div data-section="weekly-quiz">
-              <WeeklyWisdomQuiz />
-            </div>
-
             {/* Daily Quest */}
-            <div data-section="daily-quest">
-              <DailyQuest />
-            </div>
+            <DailyQuest />
             
             {/* Weekly Hackathon */}
-            <div data-section="weekly-hackathon">
-              <WeeklyHackathon />
-            </div>
+            <WeeklyHackathon />
             
             {/* Skill Graph */}
             <SkillGraph />
@@ -102,12 +80,7 @@ const Index = () => {
                   <Target className="h-4 w-4 mr-2" />
                   Generate New Quest
                 </Button> */}
-                <Button 
-                  variant="outline" 
-                  size="sm" 
-                  className="w-full justify-start"
-                  onClick={handleViewProgressReport}
-                >
+                <Button variant="outline" size="sm" className="w-full justify-start">
                   <TrendingUp className="h-4 w-4 mr-2" />
                   View Progress Report
                 </Button>
@@ -133,28 +106,6 @@ const Index = () => {
                     <span className="text-sm">{insight}</span>
                   </div>
                 ))}
-              </CardContent>
-            </Card>
-
-            {/* XP Wallet */}
-            <Card className="bg-gradient-card shadow-card hover:shadow-hover transition-all duration-300">
-              <CardHeader>
-                <CardTitle className="flex items-center gap-2">
-                  <Star className="h-5 w-5 text-yellow-400" />
-                  XP Wallet
-                </CardTitle>
-              </CardHeader>
-              <CardContent className="flex flex-col items-center space-y-3">
-                <div className="text-2xl font-bold text-blue-900">You have <span className="text-yellow-500">achieved</span> <span className="text-yellow-500">2,380 XP!</span></div>
-                <div className="italic text-blue-700 text-sm mb-2">Make your toil count!</div>
-                <Button 
-                  variant="default" 
-                  size="lg" 
-                  className="bg-yellow-400 hover:bg-yellow-500 text-white font-bold rounded-full px-8 py-2 mt-2 shadow-lg"
-                  onClick={handleRedeemXP}
-                >
-                  REDEEM XP
-                </Button>
               </CardContent>
             </Card>
